Skip adding our signature when catching up
We should only sign a single block at a certain height to prevent
leaking our private key. It is very difficult for us to know when
adding blocks given to us by other nodes if we have in fact already
signed a different block at this height. To be safe we only sign blocks
when we receive the `EXTERNALIZE` event during SCP nomination.